+++ BleachHack-Fabric-1.14/src/main/java/bleach/hack/utils/WorldRenderUtils.java
@@ -87,11 +87,11 @@ public class WorldRenderUtils {
 		int c = 0;
 		for (Entry<Enchantment, Integer> m: EnchantmentHelper.getEnchantments(item).entrySet()) {
 			String text = I18n.translate(m.getKey().getName(2).getString());
-			
+
 			if (text.isEmpty()) continue;
-			
+
 			String subText = text.substring(0, Math.min(text.length(), 2)) + m.getValue();
-			
+
 			int w1 = mc.textRenderer.getStringWidth(subText) / 2;
 			mc.textRenderer.drawWithShadow(
 					subText, -4 - w1, c*10-1,
@@ -101,13 +101,13 @@ public class WorldRenderUtils {
 		}
 
 		GL11.glScalef(0.6F, 0.6F, 0.6F);
-		
+
 		if (item.isDamageable() && item.getMaxDamage() > 0) {
 			String dur = item.getMaxDamage() - item.getDamage() + "";
-			int color = MathHelper.hsvToRgb((float) MathHelper.clamp(item.getMaxDamage() - item.getDamage() / item.getMaxDamage(), 0f, 1f) / 3.0F, 1.0F, 1.0F);
+			int color = MathHelper.hsvToRgb(MathHelper.clamp(item.getMaxDamage() - item.getDamage() / item.getMaxDamage(), 0f, 1f) / 3.0F, 1.0F, 1.0F);
 			mc.textRenderer.drawWithShadow(dur, -8 - dur.length() * 3, 15, new Color(color >> 16 & 255, color >> 8 & 255, color & 255).getRGB());
 		}
-		
+
 		glCleanup();
 	}
 
@@ -118,7 +118,7 @@ public class WorldRenderUtils {
 		GL11.glRotatef(-mc.player.yaw, 0.0F, 1.0F, 0.0F);
 		GL11.glRotatef(mc.player.pitch, 1.0F, 0.0F, 0.0F);
 		GL11.glDisable(GL11.GL_LIGHTING);
-		GL11.glDepthFunc(GL11.GL_ALWAYS); 
+		GL11.glDepthFunc(GL11.GL_ALWAYS);
 
 		GL11.glEnable(GL11.GL_BLEND);
 		GL14.glBlendFuncSeparate(GL11.GL_SRC_ALPHA, GL11.GL_ONE_MINUS_SRC_ALPHA, GL11.GL_ONE, GL11.GL_ZERO);
@@ -129,7 +129,7 @@ public class WorldRenderUtils {
 		GL11.glEnable(GL11.GL_LIGHTING);
 		GL11.glDisable(GL11.GL_BLEND);
 		GL11.glColor4f(1.0F, 1.0F, 1.0F, 1.0F);
-		GL11.glDepthFunc(GL11.GL_LEQUAL); 
+		GL11.glDepthFunc(GL11.GL_LEQUAL);
 		GL11.glPopMatrix();
 	}
 }
